move config from GitHub::DS :arrow_right: GitHub::KV
